server/requirements.txt
web3==5.28.0
Flask==2.0.3
python-dotenv==0.20.0
tqdm==4.63.1
